refactor: remove dead code from model registry
Drops normaliseOpenRouter and its types, matchSlug, and the buildPricing
wrapper — none had a caller, and two were kept alive only by their own
tests. Hoists the duplicated toSnakeCase/excludedColumns pair out of
usage.ts and model-registry.ts into queries/upsert.ts.

Also removes the cross-provider global bare-id fallback in
buildPricingFromRegistry. It resolved none of the 54 (provider, model)
pairs in use, and it let a lookup under an unmapped provider silently
borrow an unrelated provider's rate — glm-5.1 alone ranges from $0.45
to $1.40 per 1M across the providers models.dev lists. Lookups are now
provider-scoped only, so an unresolved provider yields N.A. instead of
a plausible wrong number. Two tests cover the new behaviour.

ModelSource still carries the "openrouter" member; removing it would
mean touching the model table schema, which belongs in its own change.

Closes #341
